requirements.txt
anyio==3.7.1
blinker==1.6.2
certifi==2023.7.22
click==8.1.6
Flask==2.3.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.17.3
httpx==0.24.1
hyperframe==6.0.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector-python==8.1.0
protobuf==4.21.12
sniffio==1.3.0
Werkzeug==2.3.6
